% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_specification.R
\docType{data}
\name{model_metadata}
\alias{model_metadata}
\title{Grouped List of Model Metadata}
\format{
A list containing:
\describe{
\item{N_grid}{integer: number of grid layers in which the model is divided}
\item{species}{a vector of the model-species names in the order they are taken and returned by the model}
\item{tableau_species}{a vector of the species names in the order they are returned by the tableau solving routine}
\item{reactions}{a vector of the reaction rate names in the order they are returned by the model function}
\item{omegas}{a vector of saturation indices in the order they are returned by the model function}
\item{environmental_parms}{a vector of names of environmental parameters or forcings; used for seperation in UI}
\item{boundary_conditions}{a vector of names of boundary conditions; used for seperation in UI}
\item{reaction_parms}{a vector of names of reaction parameters; used for seperation in UI}
\item{environmental_parms}{a vector of names of boundary conditions; used for seperation in UI}
}
}
\usage{
model_metadata
}
\description{
A list containing model metadata.
The entries are as follows:
}
\keyword{datasets}
