% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_state}
\alias{reference_state}
\title{Steady State Result fitted to 'greifensee_data'}
\format{
A list returned by 'solve_steady'.
}
\usage{
data(reference_state)
}
\description{
Steady State Result fitted to 'greifensee_data'
}
\keyword{datasets}
