% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTMGreifenseeApp.R
\name{RTMGreifenseeApp}
\alias{RTMGreifenseeApp}
\title{Shiny application to analyze steady-state behaviour}
\usage{
RTMGreifenseeApp()
}
\description{
Shiny application to analyze steady-state behaviour of a reactive-transport
sediment model.
}
\details{
The app initializes with measured data from Greifensee in \code{\link{greifensee_data}}
as well as fitted steady-state profiles with the tag \code{ref}.
In the sidebar parameters can be altered, a new tag can be given and the new result will
be added to a steay-state-scenarios list.
This list can be downloaded for further analysis of the scenarios.
The UI allows an upload of a stedy-state-scenarios list, which was downloaded prior.
}
